chore: clippy clean-up — auto-fix + 4 manual residuals
After enabling cargo build verification (PATH=/usr/bin:$PATH workaround
for the GCC 15 / auto-allocator issue documented in CLAUDE.md), clippy
flagged 47 unique warnings. cargo clippy --fix auto-applied 31 in one
go (mostly assert_eq!(x, true) → assert!(x) in cli_apply tests, plus
unused-variable underscore prefixes). Manual residuals (4):

- crates/treemap/src/lib.rs:957 — clippy::field_reassign_with_default
  in the SequoiaView layout test (Wave B). Switched to struct-update
  syntax (`TreeMapOptions { style: ..., ..Default::default() }`).
- src/scanner_ntfs.rs:67/623/748 — three cfg(not(windows)) stub
  functions (probe_raw_volume_access, diagnose_fsctl_enum_usn,
  mft_dump_names). They exist for API parity with the cfg(windows)
  real implementations and are never called on Linux. Annotated
  with `#[allow(dead_code)] // API-parity stub` instead of removing.

Also pulls in a Cargo.lock entry for `log` that should have shipped
with commit 8fd1f87 (render-core: log GPU adapter failures).

Verified locally with PATH=/usr/bin:$PATH:
  cargo build --workspace --all-targets       — 0 warnings, ok in 44.96s
  cargo clippy --workspace --all-targets -- -D warnings  — 0 warnings
  cargo test --workspace                      — 24 unit tests passed

+++ src/app/helpers.rs
@@ -253,8 +253,8 @@ pub(super) fn disk_free_total(path: &str) -> Option<(u64, u64)> {
         let result = unsafe { libc::statvfs(c_path.as_ptr(), stat.as_mut_ptr()) };
         if result == 0 {
             let stat = unsafe { stat.assume_init() };
-            let total = stat.f_blocks as u64 * stat.f_frsize as u64;
-            let free = stat.f_bavail as u64 * stat.f_frsize as u64;
+            let total = stat.f_blocks * stat.f_frsize;
+            let free = stat.f_bavail * stat.f_frsize;
             Some((free, total))
         } else {
             None
